+++ SceneJS/src/org/bimserver/geometry/jsonshell/SceneJs3MinimalShellSerializer.java
@@ -311,12 +311,12 @@ public class SceneJs3MinimalShellSerializer extends AbstractGeometrySerializer {
 
 	private void writeBounds(JsonWriter jsonWriter) throws IOException {
 		Extends sceneExtends = getSceneExtends();
-		float[] bounds = { (sceneExtends.max[0] + sceneExtends.min[0]) * 0.5f, (sceneExtends.max[1] + sceneExtends.min[1]) * 0.5f,
+		double[] bounds = { (sceneExtends.max[0] + sceneExtends.min[0]) * 0.5f, (sceneExtends.max[1] + sceneExtends.min[1]) * 0.5f,
 				(sceneExtends.max[2] + sceneExtends.min[2]) * 0.5f };
 		jsonWriter.beginArray();
-		jsonWriter.value(Float.isInfinite(bounds[0]) || Float.isNaN(bounds[0]) ? 50.0f : bounds[0]);
-		jsonWriter.value(Float.isInfinite(bounds[1]) || Float.isNaN(bounds[1]) ? 50.0f : bounds[1]);
-		jsonWriter.value(Float.isInfinite(bounds[2]) || Float.isNaN(bounds[2]) ? 50.0f : bounds[2]);
+		jsonWriter.value(Double.isInfinite(bounds[0]) || Double.isNaN(bounds[0]) ? 50.0f : bounds[0]);
+		jsonWriter.value(Double.isInfinite(bounds[1]) || Double.isNaN(bounds[1]) ? 50.0f : bounds[1]);
+		jsonWriter.value(Double.isInfinite(bounds[2]) || Double.isNaN(bounds[2]) ? 50.0f : bounds[2]);
 		jsonWriter.endArray();
 	}
