+++ Shared/src/org/bimserver/plugins/renderengine/RenderEngineInstance.java
@@ -18,7 +18,7 @@ package org.bimserver.plugins.renderengine;
  *****************************************************************************/
 
 public interface RenderEngineInstance {
-	float[] getTransformationMatrix() throws RenderEngineException;
+	double[] getTransformationMatrix() throws RenderEngineException;
 	RenderEngineGeometry generateGeometry() throws RenderEngineException;
 	double getArea() throws RenderEngineException;
 	double getVolume() throws RenderEngineException;
